Add support for GNU less in LeonOS with PTY and ANSI terminal adapter
- Introduced a new `storage_read_cursor` structure to manage sequential file reads in FAT32.
- Updated `storage_read_node_cursor` function to utilize the new cursor for reading files.
- Modified syscall handling to initialize and reset the read cursor for file descriptors.
- Added a build script for upstream GNU less, integrating it with LeonOS's PTY and ANSI capabilities.
- Implemented termcap functions in `leonos_termcap.c` to support terminal capabilities for less.
- Configured build-time options in `defines.h` to restrict features and ensure compatibility with LeonOS.
- Updated acknowledgements in the installer to include GNU less and Fastfetch.
- Added README documentation for GNU less usage and installation in LeonOS.
- Ensured proper integration of less with existing command resolvers and application paths.

+++ userland/fastfetch/detection_leonos.c
@@ -116,9 +116,9 @@ const char* ffDetectMemory(FFMemoryResult* result)
     if (!result || leonos_perf_info(&perf) < 0) {
         return "LeonOS performance information is unavailable";
     }
-    result->bytesTotal = perf.total_memory_kib * 1024U;
+    result->bytesTotal = (uint64_t)perf.total_memory_kib * 1024ULL;
     result->bytesUsed = perf.total_memory_kib > perf.free_memory_kib
-        ? (perf.total_memory_kib - perf.free_memory_kib) * 1024U
+        ? (uint64_t)(perf.total_memory_kib - perf.free_memory_kib) * 1024ULL
         : 0;
     return nullptr;
 }
